feat: introduce the stacks-codec component
