+++ components/clarinet-deployments/Cargo.toml
@@ -27,10 +27,11 @@ base64 = "0.21.3"
 tiny-hderive = { version = "0.3.0", optional = true }
 libsecp256k1 = { version = "0.7.0", optional = true }
 clarinet_utils = { package = "clarinet-utils", path = "../clarinet-utils", optional = true }
+stacks-codec = { path = "../stacks-codec", optional = true }
 
 [features]
 default = ["cli"]
-cli = ["clarity-repl/cli", "clarinet-files/cli", "onchain"]
+cli = ["clarity-repl/cli", "clarinet-files/cli", "stacks-codec", "onchain"]
 wasm = ["clarity-repl/wasm", "clarinet-files/wasm"]
 onchain = [
     "stacks-rpc-client",
