+++ components/clarity-repl/Cargo.toml
@@ -32,12 +32,10 @@ serde_derive = "1.0"
 integer-sqrt = "0.1.3"
 getrandom = { version = "0.2.3", features = ["js"] }
 atty = "0.2.14"
-# clarity-vm = "2"
-clarity = { git = "https://github.com/stacks-network/stacks-core.git", branch="feat/clarity-wasm-next", package = "clarity", optional = true, default-features = false }
-# clarity = { version = "2", default-features = false, optional = true, path ="../../../clarity-wasm/stacks-core/clarity" }
 clar2wasm = { git = "https://github.com/stacks-network/clarity-wasm.git", optional = true }
 # clar2wasm = { path="../../../clarity-wasm/clar2wasm", optional = true }
 pox-locking = { git = "https://github.com/stacks-network/stacks-core.git", branch="feat/clarity-wasm-next",  optional = true, default-features = false }
+stacks-codec = { path = "../stacks-codec", optional = true, default-features = false}
 
 # DAP Debugger
 tokio = { version = "1.35.1", features = ["full"], optional = true }
@@ -58,7 +56,6 @@ reqwest = { version = "0.11", default-features = false, features = [
     "json",
     "rustls-tls",
 ] }
-wsts = { version = "8.1.0", default-features = false, optional = true }
 
 # WASM
 wasm-bindgen = { version = "0.2.91", optional = true }
@@ -82,21 +79,15 @@ cli = [
     "pico-args",
     "rustyline",
     "prettytable-rs",
-    "clarity/canonical",
-    "clarity/developer-mode",
-    "clarity/log",
     "hiro_system_kit/tokio_helpers",
     "clar2wasm",
     "pox-locking/default",
-    "wsts"
+    "stacks-codec/default",
 ]
 sdk = [
-    "clarity/canonical",
-    "clarity/developer-mode",
-    "clarity/log",
     "hiro_system_kit/tokio_helpers",
     "pox-locking/default",
-    "wsts"
+    "stacks-codec/default",
 ]
 dap = [
     "tokio",
@@ -111,8 +102,7 @@ dap = [
 wasm = [
     "wasm-bindgen",
     "wasm-bindgen-futures",
-    "clarity/wasm",
-    "clarity/developer-mode",
+    "stacks-codec/wasm",
     "pox-locking/wasm"
 ]
