+++ components/stacks-codec/src/codec.rs
