+++ components/stacks-codec/src/lib.rs
@@ -1,0 +1,12 @@
+pub mod codec;
+pub mod pox;
+
+#[macro_use]
+mod macros;
+
+pub mod clarity {
+    #![allow(ambiguous_glob_reexports)]
+    pub use ::clarity::types::*;
+    pub use ::clarity::vm::*;
+    pub use ::clarity::*;
+}
