+++ components/stacks-network/Cargo.toml
@@ -45,6 +45,7 @@ clarinet-deployments = { path = "../clarinet-deployments", features = ["cli"] }
 clarinet-utils = { path = "../clarinet-utils" }
 hiro-system-kit = { path = "../hiro-system-kit", features = ["log"] }
 clarity-repl = { path = "../clarity-repl", features = ["cli"] }
+stacks-codec = { path = "../stacks-codec" }
 
 [lib]
 name = "stacks_network"
