+++ components/stacks-network/src/chains_coordinator.rs
@@ -32,10 +32,10 @@ use clarity_repl::clarity::vm::types::{BuffData, SequenceData, TupleData};
 use clarity_repl::clarity::vm::ClarityName;
 use clarity_repl::clarity::vm::Value as ClarityValue;
 use clarity_repl::clarity::PublicKey;
-use clarity_repl::codec;
 use hiro_system_kit;
 use hiro_system_kit::slog;
 use hiro_system_kit::yellow;
+use stacks_codec::codec;
 use stacks_rpc_client::PoxInfo;
 use stacks_rpc_client::StacksRpc;
 use stackslib::chainstate::stacks::address::PoxAddress;
