+++ components/stacks-rpc-client/Cargo.toml
@@ -20,6 +20,4 @@ sha2 = "0.10.0"
 tiny-hderive = { version = "0.3.0" }
 libsecp256k1 = { version = "0.7.0" }
 
-clarity_repl = { version = "2", package = "clarity-repl", path = "../clarity-repl", default-features = false, features = [
-    "sdk",
-] }
+stacks-codec = { version = "2", package = "stacks-codec", path = "../stacks-codec" }
