+++ components/stacks-rpc-client/src/crypto.rs
@@ -4,18 +4,18 @@ use crate::clarity::codec::*;
 
 use crate::clarity::vm::types::{PrincipalData, QualifiedContractIdentifier, Value};
 use crate::clarity::vm::{ClarityName, ClarityVersion, ContractName};
-use clarity_repl::clarity::address::{
-    AddressHashMode, C32_ADDRESS_VERSION_MAINNET_SINGLESIG, C32_ADDRESS_VERSION_TESTNET_SINGLESIG,
-};
-use clarity_repl::clarity::chainstate::StacksAddress;
-use clarity_repl::clarity::codec::StacksMessageCodec;
-use clarity_repl::clarity::util::secp256k1::{
-    MessageSignature, Secp256k1PrivateKey, Secp256k1PublicKey,
-};
 use hmac::Hmac;
 use libsecp256k1::{PublicKey, SecretKey};
 use pbkdf2::pbkdf2;
 use sha2::Sha512;
+use stacks_codec::clarity::address::{
+    AddressHashMode, C32_ADDRESS_VERSION_MAINNET_SINGLESIG, C32_ADDRESS_VERSION_TESTNET_SINGLESIG,
+};
+use stacks_codec::clarity::chainstate::StacksAddress;
+use stacks_codec::clarity::codec::StacksMessageCodec;
+use stacks_codec::clarity::util::secp256k1::{
+    MessageSignature, Secp256k1PrivateKey, Secp256k1PublicKey,
+};
 use tiny_hderive::bip32::ExtendedPrivKey;
 
 #[derive(Clone, Debug)]
